Move applyInput() into separate unit.
